+++ components/o-tooltip/src/js/tooltip.js
@@ -169,7 +169,7 @@ class Tooltip {
 			this.tooltipEl.style.zIndex = this.opts.zIndex;
 		}
 
-		if (this.opts.showOnClick || this.opts.toggleOnClick) {
+		if (this.opts.showOnConstruction || this.opts.showOnClick || this.opts.toggleOnClick) {
 			// Build and append the close button
 			const button = document.createElement('button');
 			button.className = 'o-tooltip-close';